+++ src/FSItem.m
@@ -311,7 +311,7 @@ static struct _BulkCatalogInfoRec {
 		return nil;
 }
 
-- (FSItem*) childAtIndex: (unsigned) index
+- (FSItem*) childAtIndex: (NSUInteger) index
 {
 	if ( ![self isSpecialItem] )
 		return [_childs objectAtIndex: index];
@@ -319,7 +319,7 @@ static struct _BulkCatalogInfoRec {
 		return nil;
 }
 
-- (unsigned) childCount
+- (NSUInteger) childCount
 {
 	if ( ![self isSpecialItem] )
 		return [_childs count];
@@ -433,7 +433,7 @@ static struct _BulkCatalogInfoRec {
 		case FileFolderItem:
 			if ( [self isFolder] )
 			{
-				unsigned i = [_childs count];
+				NSUInteger i = [_childs count];
 				while ( i-- )
 				{
 					FSItem *child = [_childs objectAtIndex: i];
@@ -594,7 +594,7 @@ static struct _BulkCatalogInfoRec {
 	//let our childs do the same
 	if ( includeChilds && [self isFolder] )
 	{
-		unsigned i = [self childCount];
+		NSUInteger i = [self childCount];
 		while ( i-- )
 			[[self childAtIndex: i] setKindStringIgnoringCreatorCode: ignoreCreatorCode includeChilds: YES];
 	}
